MXS-2245: Add main replication class
The Replicator class handles the replication of events from the master and
the distribution of row events to the individual tables. It also acts as a
coordinator that drives the conversion process by synchronizing the
threads between transitions from row event processing into query event
processing.

The cdc::Replicator class in replicator.hh is currently the only
public-facing API to the implementation. It is unlikely that the code will
be used as a full-blown library so the "hiding" of the implementation
details acts more as an exercise in API/ABI cleanliness.