MXS-2245: Add high-level event processing
GTID events are handled at this level as they are used by the Replicator
when a reconnection occurs. The actual processing of events into bulk
inserts and SQL statements still remains to be implemented. With these
changes the Replicator should now be able to handle master failures as
long as the master replacements have the GTIDs in their binlogs.