MXS-2245: Add initial query execution
The SQLExecutor class executes SQL queries asynchronously on a set of
servers. This allows the replicating thread to read more events while the
SQL thread executes them. In addition to executing the events on the
ColumnStore server, any tables created with an implicit default storage
engine are converted to use the ColumnStore storage engine.