MXS-2245: Make event processing transactional
A failure to process an event causes the processing to stop in a
controlled manner. By making the event processing transactional, only
complete transactions are processed.

There are potential problems where the committing of a transaction appears
to fail but in reality the transaction gets committed. Upon restarting,
the processing would continue from the last known good GTID position which
would be the one before the failed transaction. This would cause the
transaction to appear twice in ColumnStore possibly causing further
problems. A solution to this would be to use XA transactions for query
events but for bulk inserts the problem still remains.