MXS-2245: Process replicated events as transactions
By flushing all pending changes when a transaction is committed, we will
always know the last committed GTID position. This prevents the case where
multiple transactions are queued to a REProc and the first one of them
fails.

The exception to this is the bulk loading. Since we know that all events
add new data, multiple transactions can be processed without committing
between them. This retains the speed advantage of using the bulk loading
API. Currently the code flushes the bulk loads periodically or whenever a
non-INSERT row event is received.