MXS-2245: Join Replicator thread on exit
When the Replicator class was destoyed, the thread would be orphaned
causing it throw an exception. To prevent this, the stopping logic needs
to be chaned so that the stop method only starts the shutdown process but
does not wait for it to complete. The synchronization is done when the
Replictor class is destroyed which keeps it simple.