MXS-2245: Add transformation mode
The transformation mode transforms updates and deletes into inserts. This
makes it possible to store a history of changes into ColumnStore similar
to what the ColumnStore mxs_adapter data adapter originally did.

The code is not very useful in its current form as there is no way to
distinct inserts, updates and deletes from each other. Additional
information needs to be stored in each row to make it possible to impose
some sort of ordering between the rows as well as being able to easily
distinct deleted rows from inserted rows.