MXS-2245: Start replication from FDE
The FDE at the start of the binlog must be read to know whether binlog
checksums are enabled. As the replication can stop mid-binlog, the
replication has to always start at the beginning of the binlog file that
contains the target GTID where the replication should continue. If the
target GTID is the last GTID in the file, the next file is used.

Since the events must not be processed twice, any events that happened
before the target GTID position must be ignored. Once the target GTID is
encountered, the next transaction is ignored after which processing
continues normally.

This change depends on pull request #100:
https://github.com/MariaDB/mariadb-connector-c/pull/100