MXS-2245: Add row event to SQL transformation
The row events can be transformed into simple SQL statements to update and
delete data on ColumnStore. This is logically close to statement based
replication but it is significantly less efficient, especially so with how
ColumnStore and non-insert DMLs work.

The use of the LIMIT 1 clause prevents updating and deletion of multiple
matching rows with a single SQL statement. This preserves the data but
mutates the unsorted order of the rows. As ColumnStore does not have a
primary key for each row, it is impossible to distinct rows from each
other thus preventing updates that preserve the unsorted order.