MXS-2245: Take SQL transformation code into use
The update and delete rows events are now converted into sets of SQL
statements that are executed on ColumnStore via the UM. If a transaction
mixes inserts, updates and deletes the inserts are pre-emptively committed
on the database. This is done as a bulk insert locks the table preventing
further updates to it before the bulk insert is committed.

Also added the missing BEGIN and COMMIT statements to the SQL command
batches.