MXS-2245: Avoid stopping mid-transaction
If the process shutdown is started when the replication stream is in the
middle of a transaction, the Replicator attemps to continue processing
until a safe stopping point is found. Currently, these safe stopping
points are the start of a new tansaction, the ending of a finished
transaction as well as the rotation to a new binlog file. In theory the
processing could be stopped in other places as well but these are not very
common.