Prototype UNIX domain socket support for servers
The `address` parameter of a server now accepts absolute paths to UNIX
domain sockets. The code only checks for the first character which is why
relative paths are not supported.

Due to the way the mariadbmon checks for the replication topology in 2.3,
the assume_unique_hostnames=false parameter must be added to the
monitor. This also means that automatic failover is not supported as due
to the requirement for unique hostnames.